门店管理

+++ mallplus-admin/src/main/java/com/zscat/mallplus/config/MybatisPlusConfig.java
@@ -53,14 +53,14 @@ public class MybatisPlusConfig {
             @Override
             public Expression getTenantId() {
                 // 从当前系统上下文中取出当前请求的服务商ID，通过解析器注入到SQL中。
-                Long currentProviderId = Long.valueOf(UserUtils.getCurrentMember().getStoreId());
+                Integer currentProviderId = UserUtils.getCurrentMember().getStoreId();
                 System.out.println("currentProviderId=" + currentProviderId);
                 if (null == currentProviderId) {
-                    currentProviderId = 1l;
+                    currentProviderId = 1;
                     System.out.println("#1129 getCurrentProviderId error.");
                     //  throw new RuntimeException("#1129 getCurrentProviderId error.");
                 }
-                return new LongValue(currentProviderId);
+                return new LongValue(Long.valueOf(currentProviderId));
             }
 
             @Override
